# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Participant {
  id: string;
  name: string;
  role: string;
  joinedAt: string;
  isMuted: boolean;
  cameraOn: boolean;
}

export interface TranscriptEntry {
  id: string;
  speaker: string;
  text: string;
  createdAt: string;
}

export interface MeetingNotes {
  summary: string;
  keyPoints: string[];
  decisions: string[];
  actionItems: string[];
  followUps: string[];
  generatedAt: string;
}

export type MeetingStatus = typeof MeetingStatus[keyof typeof MeetingStatus];


export const MeetingStatus = {
  live: 'live',
  scheduled: 'scheduled',
  ended: 'ended',
} as const;

export interface Meeting {
  id: string;
  title: string;
  hostId: string;
  hostName: string;
  createdAt: string;
  startedAt: string;
  /** @nullable */
  endedAt: string | null;
  status: MeetingStatus;
  participants: Participant[];
  transcript: TranscriptEntry[];
  notes: MeetingNotes | null;
  participantCount: number;
}

export interface MeetingInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  hostName: string;
}

export type MeetingUpdateStatus = typeof MeetingUpdateStatus[keyof typeof MeetingUpdateStatus];


export const MeetingUpdateStatus = {
  live: 'live',
  scheduled: 'scheduled',
  ended: 'ended',
} as const;

export interface MeetingUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title?: string;
  status?: MeetingUpdateStatus;
  /** @nullable */
  endedAt?: string | null;
}

export interface TranscriptInput {
  /** @minLength 1 */
  speaker: string;
  /** @minLength 1 */
  text: string;
}

export interface Dashboard {
  totalMeetings: number;
  liveMeetings: number;
  summaries: number;
  actionItems: number;
  hoursSaved: number;
  recent: Meeting[];
  upcoming: Meeting[];
}

export type ListMeetingsParams = {
status?: ListMeetingsStatus;
};

export type ListMeetingsStatus = typeof ListMeetingsStatus[keyof typeof ListMeetingsStatus];


export const ListMeetingsStatus = {
  live: 'live',
  scheduled: 'scheduled',
  ended: 'ended',
  all: 'all',
} as const;