# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { Meeting } from './meeting';

export interface Dashboard {
  totalMeetings: number;
  liveMeetings: number;
  summaries: number;
  actionItems: number;
  hoursSaved: number;
  recent: Meeting[];
  upcoming: Meeting[];
}