# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { MeetingNotes } from './meetingNotes';
import type { MeetingStatus } from './meetingStatus';
import type { Participant } from './participant';
import type { TranscriptEntry } from './transcriptEntry';

export interface Meeting {
  id: string;
  title: string;
  hostId: string;
  hostName: string;
  createdAt: Date;
  startedAt: Date;
  /** @nullable */
  endedAt: Date | null;
  status: MeetingStatus;
  participants: Participant[];
  transcript: TranscriptEntry[];
  notes: MeetingNotes | null;
  participantCount: number;
}