# lib/api-zod/src/generated/types/meetingNotes.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingNotes {
  summary: string;
  keyPoints: string[];
  decisions: string[];
  actionItems: string[];
  followUps: string[];
  generatedAt: Date;
}