# lib/api-zod/src/generated/types/meetingUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { MeetingUpdateStatus } from './meetingUpdateStatus';

export interface MeetingUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title?: string;
  status?: MeetingUpdateStatus;
  /** @nullable */
  endedAt?: Date | null;
}