# lib/api-zod/src/generated/types/participant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Participant {
  id: string;
  name: string;
  role: string;
  joinedAt: Date;
  isMuted: boolean;
  cameraOn: boolean;
}