# lib/api-zod/src/generated/types/transcriptEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptEntry {
  id: string;
  speaker: string;
  text: string;
  createdAt: Date;
}