# lib/api-zod/src/generated/types/transcriptInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gen Z Meet meeting workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptInput {
  /** @minLength 1 */
  speaker: string;
  /** @minLength 1 */
  text: string;
}